+++ eras/conway/test-suite/src/Test/Cardano/Ledger/Conway/Translation/TranslatableGen.hs
@@ -63,3 +63,4 @@ genTxBody l = do
     <*> arbitrary
     <*> arbitrary
     <*> arbitrary
+    <*> arbitrary