+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState/NewEpochState.hs
@@ -91,6 +91,7 @@ genesisState genDelegs0 utxo0 =
         (Coin 0)
         emptyGovState
         (IStake mempty Map.empty)
+        mempty
     )
     (CertState def def dState)
   where