+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/AlonzoValidTxUTXOW.hs
@@ -51,7 +51,7 @@ import Cardano.Ledger.Shelley.TxBody (
  )
 import Cardano.Ledger.Shelley.TxCert (pattern UnRegTxCert)
 import Cardano.Ledger.TxIn (TxIn (..))
-import Cardano.Ledger.Val (inject, (<+>))
+import Cardano.Ledger.Val (Val (..), inject, (<+>))
 import Cardano.Slotting.Slot (SlotNo (..))
 import Control.State.Transition.Extended hiding (Assertion)
 import Data.Default.Class (Default (..))
@@ -231,7 +231,7 @@ validatingState ::
   ) =>
   Proof era ->
   UTxOState era
-validatingState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+validatingState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def mempty
   where
     utxo = expectedUTxO' pf (ExpectSuccess (validatingBody pf) (validatingTxOut pf)) 1
 
@@ -286,7 +286,7 @@ notValidatingState ::
   ) =>
   Proof era ->
   UTxOState era
-notValidatingState pf = smartUTxOState (pp pf) (expectedUTxO' pf ExpectFailure 2) (Coin 0) (Coin 5) def
+notValidatingState pf = smartUTxOState (pp pf) (expectedUTxO' pf ExpectFailure 2) (Coin 0) (Coin 5) def mempty
 
 -- =========================================================================
 --  Example 3: Process a CERT transaction with a succeeding Plutus script.
@@ -340,7 +340,7 @@ validatingWithCertState ::
   ) =>
   Proof era ->
   UTxOState era
-validatingWithCertState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+validatingWithCertState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def mempty
   where
     utxo = expectedUTxO' pf (ExpectSuccess (validatingWithCertBody pf) (validatingWithCertTxOut pf)) 3
 
@@ -389,7 +389,14 @@ notValidatingWithCertState ::
   ) =>
   Proof era ->
   UTxOState era
-notValidatingWithCertState pf = smartUTxOState (pp pf) (expectedUTxO' pf ExpectFailure 4) (Coin 0) (Coin 5) def
+notValidatingWithCertState pf =
+  smartUTxOState
+    (pp pf)
+    (expectedUTxO' pf ExpectFailure 4)
+    (Coin 0)
+    (Coin 5)
+    def
+    mempty
 
 -- ==============================================================================
 --  Example 5: Process a WITHDRAWAL transaction with a succeeding Plutus script.
@@ -443,7 +450,14 @@ validatingWithWithdrawalState ::
   (EraTxBody era, PostShelley era, EraGov era) =>
   Proof era ->
   UTxOState era
-validatingWithWithdrawalState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+validatingWithWithdrawalState pf =
+  smartUTxOState
+    (pp pf)
+    utxo
+    (Coin 0)
+    (Coin 5)
+    def
+    mempty
   where
     utxo = expectedUTxO' pf (ExpectSuccess (validatingWithWithdrawalBody pf) (validatingWithWithdrawalTxOut pf)) 5
 
@@ -492,7 +506,14 @@ notValidatingWithWithdrawalState ::
   (EraTxBody era, PostShelley era, EraGov era) =>
   Proof era ->
   UTxOState era
-notValidatingWithWithdrawalState pf = smartUTxOState (pp pf) (expectedUTxO' pf ExpectFailure 6) (Coin 0) (Coin 5) def
+notValidatingWithWithdrawalState pf =
+  smartUTxOState
+    (pp pf)
+    (expectedUTxO' pf ExpectFailure 6)
+    (Coin 0)
+    (Coin 5)
+    def
+    mempty
 
 -- =============================================================================
 --  Example 7: Process a MINT transaction with a succeeding Plutus script.
@@ -550,7 +571,14 @@ validatingWithMintState ::
   (PostShelley era, EraTxBody era, HasTokens era, Value era ~ MaryValue (EraCrypto era), EraGov era) =>
   Proof era ->
   UTxOState era
-validatingWithMintState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+validatingWithMintState pf =
+  smartUTxOState
+    (pp pf)
+    utxo
+    (Coin 0)
+    (Coin 5)
+    def
+    mempty
   where
     utxo = expectedUTxO' pf (ExpectSuccess (validatingWithMintBody pf) (validatingWithMintTxOut pf)) 7
 
@@ -598,7 +626,7 @@ notValidatingWithMintState ::
   (EraTxBody era, PostShelley era, EraGov era) =>
   Proof era ->
   UTxOState era
-notValidatingWithMintState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+notValidatingWithMintState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def zero
   where
     utxo = expectedUTxO' pf ExpectFailure 8
 
@@ -688,10 +716,23 @@ validatingManyScriptsTxOut pf =
 
 validatingManyScriptsState ::
   forall era.
-  (EraTxBody era, PostShelley era, HasTokens era, Value era ~ MaryValue (EraCrypto era), EraGov era, ShelleyEraTxCert era) =>
+  ( EraTxBody era
+  , PostShelley era
+  , HasTokens era
+  , Value era ~ MaryValue (EraCrypto era)
+  , EraGov era
+  , ShelleyEraTxCert era
+  ) =>
   Proof era ->
   UTxOState era
-validatingManyScriptsState pf = smartUTxOState (pp pf) (UTxO utxo) (Coin 0) (Coin 5) def
+validatingManyScriptsState pf =
+  smartUTxOState
+    (pp pf)
+    (UTxO utxo)
+    (Coin 0)
+    (Coin 5)
+    def
+    mempty
   where
     utxo =
       Map.insert (TxIn (txid (validatingManyScriptsBody pf)) minBound) (validatingManyScriptsTxOut pf) $
@@ -748,7 +789,8 @@ validatingSupplimentaryDatumState ::
   (EraTxBody era, PostShelley era, EraGov era) =>
   Proof era ->
   UTxOState era
-validatingSupplimentaryDatumState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+validatingSupplimentaryDatumState pf =
+  smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def mempty
   where
     utxo = expectedUTxO' pf (ExpectSuccess (validatingSupplimentaryDatumBody pf) (validatingSupplimentaryDatumTxOut pf)) 3
 
@@ -805,7 +847,8 @@ validatingMultipleEqualCertsState ::
   (EraTxBody era, PostShelley era, EraGov era, ShelleyEraTxCert era) =>
   Proof era ->
   UTxOState era
-validatingMultipleEqualCertsState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+validatingMultipleEqualCertsState pf =
+  smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def mempty
   where
     utxo = expectedUTxO' pf (ExpectSuccess (validatingMultipleEqualCertsBody pf) (validatingMultipleEqualCertsOut pf)) 3
 
@@ -853,7 +896,8 @@ validatingNonScriptOutWithDatumState ::
   ) =>
   Proof era ->
   UTxOState era
-validatingNonScriptOutWithDatumState pf = smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def
+validatingNonScriptOutWithDatumState pf =
+  smartUTxOState (pp pf) utxo (Coin 0) (Coin 5) def mempty
   where
     utxo = expectedUTxO' pf (ExpectSuccess (validatingNonScriptOutWithDatumTxBody pf) (validatingNonScriptOutWithDatumTxOut pf)) 103