+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Fields.hs
@@ -126,6 +126,7 @@ data TxBodyField era
   | Txnetworkid (StrictMaybe Network)
   | GovProcs (GovProcedures era)
   | CurrentTreasuryValue (StrictMaybe Coin)
+  | TreasuryDonation Coin
 
 pattern Inputs' :: [TxIn (EraCrypto era)] -> TxBodyField era -- Set
 
@@ -312,7 +313,7 @@ abstractTxBody (Alonzo _) (AlonzoTxBody inp col out cert wdrl fee vldt up req mn
   , AdHash adh
   , Txnetworkid net
   ]
-abstractTxBody (Conway _) (ConwayTxBody inp col ref out colret totcol cert wdrl fee vldt req mnt sih adh net vp pp ctv) =
+abstractTxBody (Conway _) (ConwayTxBody inp col ref out colret totcol cert wdrl fee vldt req mnt sih adh net vp pp ctv td) =
   [ Inputs inp
   , Collateral col
   , RefInputs ref
@@ -330,6 +331,7 @@ abstractTxBody (Conway _) (ConwayTxBody inp col ref out colret totcol cert wdrl 
   , Txnetworkid net
   , GovProcs $ GovProcedures vp (fromStrict pp)
   , CurrentTreasuryValue ctv
+  , TreasuryDonation td
   ]
 abstractTxBody (Babbage _) (BabbageTxBody inp col ref out colret totcol cert wdrl fee vldt up req mnt sih adh net) =
   [ Inputs inp